more refactoring of Player class
